+++ forge/src/main/java/net/mehvahdjukaar/every_compat/forge/ECPlatformStuffImpl.java
@@ -2,7 +2,7 @@ package net.mehvahdjukaar.every_compat.forge;
 
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.ToolActions;
-import org.violetmoon.quark.base.handler.ToolInteractionHandler;
+import org.violetmoon.zeta.util.handler.ToolInteractionHandler;
 
 public class ECPlatformStuffImpl {